% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{add.visit.name}
\alias{add.visit.name}
\title{add visit names to a table.}
\usage{
add.visit.name(data)
}
\arguments{
\item{data}{data.frame to add visit name to}
}
\value{
data.frame with additional mnpvislabel variable for visit label
}
\description{
Add the visit name to a table (as opposed to just the visit id - secuTrial uses many visit ids to refer to a single visit).
Uses results of \code{load.study.options} directly - must be run after \code{load.tables} or \code{load.study.options}
}
\examples{
# TODO!!
}
\references{
http://stackoverflow.com/questions/3640925/global-variable-in-r-function
}
\seealso{
load.table, load.study.options
}
