% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dates.R
\name{convert.all.dates}
\alias{convert.all.dates}
\title{Convert all date columns of a data.frame to Date vectors.}
\usage{
convert.all.dates(dat, .format = "\%Y-\%m-\%d",
  convert.unknown.date.to.na = FALSE, unknown.date.string = NULL,
  partial.date.handling = "skip", partial.date.string = "Unknown",
  silent = FALSE)
}
\arguments{
\item{dat}{data frame to convert dates}

\item{.format}{format of the dates (default: '\%Y-\%m-\%d')}

\item{convert.unknown.date.to.na}{flag for conversion of unknown dates}

\item{unknown.date.string}{string specifying the coding of unknown dates}

\item{partial.date.handling}{how to handle columns containing entries that cannot be converted (default: skip; see Details)}

\item{partial.date.string}{string used for unknown data components (e.g. 'Unknown' or '--')}

\item{silent}{hide verbose output}
}
\value{
Date frame with date columns converted.
}
\description{
This function identifies all date columns in a data frame and converts them
to vectors of Date objects. Optionally, unknown dates (e.g. coded as '1900-01-01')
can be converted to \code{NA}.
}
\details{
\code{partial.date.handling}: By default, potential data columns containing entries that cannot be converted are skipped (option 'skip'). With 'force.conversion' fields incompatible with the date format are converted to NA. 'fill.partial.dates' keeps the orginal column untouched and adds a new column to the data frame with the name <colname>.partial.dates.processed in which partial are filled using \code{\link{fill.partial.date}} (e.g. Unknown.01.2013 -> 15.01.2013). With 'fill.partial.dates.and.keep.original' partial dates are processed directly and the original data is copied to <colname>.original.
}
\examples{
convert.all.dates(data.frame(date=c("2014-01-Unknown", "1900-01-01")),
partial.date.handling="fill.partial.dates", convert.unknown.date.to.na = TRUE,
unknown.date.string="1900-01-01")
##              date date.partial.dates.processed
##1  2014-01-Unknown                   2014-01-15
##2       1900-01-01                         <NA>
}
\seealso{
fill.partial.date
}
\author{
Pascal Benkert
}
