% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dates.R
\name{days.of.month}
\alias{days.of.month}
\title{Number of days of given month.}
\usage{
days.of.month(date)
}
\arguments{
\item{date}{Date object}
}
\value{
Number of days.
}
\description{
This function returns the number of days for a given month as specified by
as Date object generated by \code{\link{as.Date}}.
}
\examples{
days.of.month(as.Date("01.02.2013", "\%d.\%m.\%Y")) # 28
}
\author{
Pascal Benkert
}
