% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dates.R
\name{fill.partial.date}
\alias{fill.partial.date}
\title{Fill partial dates.}
\usage{
fill.partial.date(date.string, partial.date.string = "Unknown",
  .format = "\%Y-\%m-\%d", exact = FALSE)
}
\arguments{
\item{date.string}{string containing a partial date}

\item{partial.date.string}{string used for unknown data components (e.g. 'Unknown' or '--')}

\item{.format}{format of the dates (default: '\%Y-\%m-\%d')}

\item{exact}{use the exact modpoints \code{floor(ndays/2)} and 02.07.YYYY}
}
\value{
Date string, converted if applicable.
}
\description{
This function convert a partial date of the form Unknown.mm.YYYY or
Unknown.Unknown.YYYY to the typical placeholders 15.mm.YYYY and 01.07.YYYY,
respectively. Other \code{partial.date.string} can be specified.
Alternatively, the 'exact' midpoint of the month/year can be used, i.e.
\code{floor(ndays/2)} and 02.07.YYYY.
}
\examples{
fill.partial.date("Unknown.02.2014", "Unknown", .format="\%d.\%m.\%Y", exact=TRUE) # 14.02.2014
fill.partial.date("2014/--/--", "--", .format="\%Y/\%m/\%d") # 2014/07/01
}
\author{
Pascal Benkert
}
