% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load.labels}
\alias{load.labels}
\title{Load labels from an export .}
\usage{
load.labels()
}
\value{
(Pre-processed) tables in \code{tables} as data frames
}
\description{
Get a named vector of variable labels.
Uses results of \code{load.study.options} directly - must be run after \code{load.tables} or \code{load.study.options}
}
\examples{
## non rectangular table
load.study.options(data.dir=system.file("extdata", "s_export_CSV-xls_DEM00_20180912-125720.zip", package = "secuTrial"))
labs <- load.tables()
labs[1]
}
\references{
http://stackoverflow.com/questions/3640925/global-variable-in-r-function
}
\seealso{
read.DB.table, load.table.list (used in dossier-specific packages), load.study.options
}
