% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load.tables}
\alias{load.tables}
\title{Load tables from an export and distinguish between rectangular and non-rectangular input.}
\usage{
load.tables(data.dir, tables = "all", convert.dates = FALSE,
  convert.unknown.date.to.na = FALSE, rename.headers = FALSE,
  add.pat.id = TRUE, add.center = FALSE, silent = FALSE,
  is.rt = FALSE, decode.rt.visitlabels = TRUE)
}
\arguments{
\item{data.dir}{location of the export directory (or zip file) containing the csv/xls files}

\item{tables}{vector of tables to be loaded (if a \code{table.list} exists, the corresponing table names can be given)}

\item{convert.dates}{identify and convert date columns}

\item{convert.unknown.date.to.na}{convert unknown date strings (e.g 1900-01-01) to NA}

\item{rename.headers}{rename column names}

\item{add.pat.id}{adding study ID pat.id to the first column}

\item{add.center}{adding center information}

\item{silent}{hide output}

\item{is.rt}{specifies if export is in rectangular table format}

\item{decode.rt.visitlabels}{specifies whether to decode the visit labels in the rectangular table}
}
\value{
(Pre-processed) tables in \code{tables} as data frames
}
\description{
If \code{is.rt} is TRUE then only the \code{decode.rt.visitlabels} and \code{silent} parameters are used.
If \code{is.rt} is FALSE then all other parameters including \code{silent} are interpreted.
The argument \code{tables} can handle four distinguished cases:
\itemize{
\item tables=NULL: there is variable called 'table.list' e.g. created from dossier-specific
package and all tables should be loaded. Definition:
table.list <- list(data.frame.name=list(filename=<filename in the export>,
tablename=<internal table name>), ...). The latter is only relevant for queries.
\item tables=c(tablename1,tablename2,...): there is a 'table.list', but only some
tables should be loaded.
\item tables=c(filename1,filename2,...): the user specifies the files that should be loaded.
\item tables="all" loads all files in the zip-archive or directory.
}
The optional pre-processing steps are: identification of date columns
and convertion to objects of the class 'Date', conversion of unknown dates to NA and
renaming of column names (remove '_' and capital letters, see \code{\link{new.names}}). The study specific date formats and unknown date strings
are specified in \code{load.study.options}.
}
\examples{
## non rectangular table
load.tables(data.dir=system.file("extdata", "s_export_CSV-xls_DEM00_20180912-125720.zip", package = "secuTrial"))
## rectangular table
load.tables(system.file("extdata", "s_export_rt-CSV-xls_DEM00_20181016-151332.zip", package = "secuTrial"),
            is.rt = TRUE, decode.rt.visitlabels = TRUE)
}
\references{
http://stackoverflow.com/questions/3640925/global-variable-in-r-function
}
\seealso{
read.DB.table, load.table.list (used in dossier-specific packages), load.study.options
}
