% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforming.R
\name{trim}
\alias{trim}
\title{This function trims whitespaces from a string.}
\usage{
trim(x, all = FALSE, ...)
}
\arguments{
\item{all}{will remove all whitespaces from string if set to TRUE. If FALSE only leading and trailing whitespaces will be removed.}
}
\value{
string
}
\description{
Taken from the lava package
}
\examples{
trim("  test string trim  ")
# "test string trim"
trim("  test string trim  ", all=TRUE)
# "teststringtrim"
}
